requirements.txt
Appium-Python-Client
page-objects
behave
selenium==3.3.1
behave-teamcity